% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dots_used}
\alias{check_dots_used}
\title{Check that all dots in current environment have been used}
\usage{
check_dots_used(env = caller_env())
}
\arguments{
\item{env}{Environment in which to look for \code{...}. This is a temporary
API until we have a C API that can follow promises up multiple levels
in the call stack.}
}
\description{
Typically used in \code{on.exit()}. Generates warning if any dots have not
been forced (i.e. evaluated)
}
\examples{
f <- function(...) {
  on.exit(check_dots_used())
  g(...)
}

g <- function(x, y, ...) {
  x + y
}
f(x = 1, y = 2)

f(x = 1, y = 2, z = 3)
f(x = 1, y = 2, 3, 4, 5)
}
