% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{conditions}
\alias{conditions}
\alias{stop_dots_used}
\alias{stop_dots_unnamed}
\alias{stop_dots_empty}
\title{Error conditions}
\usage{
stop_dots_used(unused_names, ..., message = NULL, .subclass = NULL)

stop_dots_unnamed(named_names, ..., message = NULL, .subclass = NULL)

stop_dots_empty(names, ..., message = NULL, .subclass = NULL)
}
\arguments{
\item{unused_names}{Names of unused dots.}

\item{...}{Additional data to be stored in the condition object.}

\item{message}{The message to display.}

\item{.subclass}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}

\item{named_names}{Names of dots that should have not been named.}

\item{names}{Names of dots.}
}
\description{
Error conditions
}
\keyword{internal}
