% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_used}
\alias{dots_used}
\title{Helper for consistent documentation of used dots}
\arguments{
\item{...}{Arguments passed to methods.}
}
\description{
Use \verb{@inheritParams ellipsis::dots_used} in your package
to consistently document an unused \code{...} argument.
}
\keyword{internal}
