% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsis-package.R
\docType{package}
\name{ellipsis-package}
\alias{ellipsis}
\alias{ellipsis-package}
\title{ellipsis: Tools for Working with ...}
\description{
In S3 generics, it's useful to take ... so that methods can
    have additional argument. But this flexibility comes at a cost: misspelled
    arguments will be silently ignored. The ellipsis packages is an experiment
    that allows a generic to warn if any arguments passed in ... are not
    used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/ellipsis}
  \item Report bugs at \url{https://github.com/r-lib/ellipsis/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
