% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.R
\name{safe}
\alias{safe}
\alias{safe_map}
\alias{safe_median}
\title{Safer version of risky functions}
\usage{
safe_map(x, f, ...)

safe_median(x, ...)
}
\arguments{
\item{x}{Vector input}

\item{f}{Function to apply to each element}

\item{...}{For \code{safe_map()}, additional arguments passed on to \code{f}.
For \code{safe_median()}, additional arguments passed on to methods.}
}
\description{
These functions warn if some elements of \code{...} never get used.
}
\examples{
x <- safe_map(iris[1:4], median, na.rm = TRUE)
x <- safe_map(iris[1:4], median, na.mr = TRUE)
}
