% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{stop_dots_not_empty}
\alias{stop_dots_not_empty}
\title{Stop with custom conditions}
\usage{
stop_dots_not_empty(message = NULL, .subclass = NULL, ...)
}
\arguments{
\item{message}{The message to display.}

\item{.subclass}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}

\item{...}{Additional data to be stored in the condition object.}
}
\description{
Use these \code{stop_} functions when you need a different message or
error class.
}
