% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dots_named}
\alias{check_dots_named}
\title{Check that dots are either empty or all dot parameter names are a valid subset of a function's parameter names.}
\usage{
check_dots_named(
  ...,
  .function,
  .forbidden = NULL,
  .empty_ok = TRUE,
  .action = abort
)
}
\arguments{
\item{...}{The dots argument to check.}

\item{.function}{The function name.}

\item{.forbidden}{Parameter names within \code{...} that should be treated as invalid.}

\item{.empty_ok}{Set to \code{TRUE} if empty \code{...} should be allowed, otherwise to \code{FALSE}.}

\item{.action}{The action to take when the dots have not been used. One of
\code{\link[rlang:abort]{rlang::abort()}}, \code{\link[rlang:warn]{rlang::warn()}}, \code{\link[rlang:inform]{rlang::inform()}} or \code{\link[rlang:signal]{rlang::signal()}}.}
}
\description{
Check that dots are either empty or all dot parameter names are a valid subset of a function's parameter names.
}
\examples{
sapply_safe <- function(X,
                        FUN,
                        ...,
                        simplify = TRUE,
                        USE.NAMES = TRUE) {
  check_dots_named(...,
                   .function = FUN)
  sapply(X = X,
         FUN = FUN,
         ...,
         simplify = TRUE,
         USE.NAMES = TRUE)
}
}
