% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_empty}
\alias{dots_empty}
\title{Helper for consistent documentation of empty dots}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Use \verb{@inheritParams ellipsis::dots_empty} in your package
to consistently document an unused \code{...} argument.
}
\keyword{internal}
