% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsis-package.R
\docType{package}
\name{ellipsis-package}
\alias{ellipsis}
\alias{ellipsis-package}
\title{ellipsis: Tools for Working with ...}
\description{
... is a powerful tool for function extensibility. Unfortunately 
    this power comes at a cost: misspelled arguments will be silently ignored. 
    The ellipsis package provides a collection of functions to catch problems
    and alert the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/ellipsis}
  \item Report bugs at \url{https://github.com/r-lib/ellipsis/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
